deleting chatIF making 1 client controller
