--- src/client/ChatIF.java
@@ -1,24 +1,0 @@
-// This file contains material supporting section 3.7 of the textbook:
-// "Object Oriented Software Engineering" and is issued under the open-source
-// license found at www.lloseng.com 
-
-package client;
-
-import java.sql.ResultSet;
-
-/**
- * This interface implements the abstract method used to display
- * objects onto the client or server UIs.
- *
- * @author Dr Robert Lagani&egrave;re
- * @author Dr Timothy C. Lethbridge
- * @version July 2000
- */
-public interface ChatIF 
-{
-  /**
-   * Method that when overriden is used to display objects onto
-   * a UI.
-   */
-  public abstract void display(String obj);
-}
